+++ pipeline/tick/alert_test.go
@@ -313,7 +313,7 @@ func TestAlertPagerDuty(t *testing.T) {
 func TestAlertPagerDuty2(t *testing.T) {
 	pipe, _, from := StreamFrom()
 	handler := from.Alert().PagerDuty2()
-	handler.ServiceKey = "LeafsNation"
+	handler.RoutingKey = "LeafsNation"
 
 	want := `stream
     |from()
@@ -323,7 +323,25 @@ func TestAlertPagerDuty2(t *testing.T) {
         .details('{{ json . }}')
         .history(21)
         .pagerDuty2()
-        .serviceKey('LeafsNation')
+        .routingKey('LeafsNation')
+`
+	PipelineTickTestHelper(t, pipe, want)
+}
+
+func TestAlertPagerDuty2_ServiceKey(t *testing.T) {
+	pipe, _, from := StreamFrom()
+	handler := from.Alert().PagerDuty2()
+	handler.ServiceKey("LeafsNation")
+
+	want := `stream
+    |from()
+    |alert()
+        .id('{{ .Name }}:{{ .Group }}')
+        .message('{{ .ID }} is {{ .Level }}')
+        .details('{{ json . }}')
+        .history(21)
+        .pagerDuty2()
+        .routingKey('LeafsNation')
 `
 	PipelineTickTestHelper(t, pipe, want)
 }